use build constraints for windows and unix
