+++ runner/go.mod
@@ -5,6 +5,7 @@ go 1.18
 require (
 	cloud.google.com/go/bigquery v1.8.0
 	github.com/ClickHouse/clickhouse-go/v2 v2.0.14
+	github.com/alexbrainman/odbc v0.0.0-20211220213544-9c9a2e61c5e2
 	github.com/aws/aws-sdk-go v1.44.14
 	github.com/denisenkom/go-mssqldb v0.12.0
 	github.com/flosch/pongo2 v0.0.0-20200913210552-0d938eb266f3
