+++ runner/odbc_windows.go
@@ -1,0 +1,10 @@
+package runner
+
+import (
+	_ "github.com/alexbrainman/odbc"
+	"github.com/jmoiron/sqlx"
+)
+
+func openODBCDriver(conn string) (*sqlx.DB, error) {
+	return sqlx.Open("odbc", conn)
+}
